requirements_airflow.txt
apache-airflow
apache-airflow-providers-apache-spark
apache-airflow-providers-cncf-kubernetes
numpy